+++ src/main/java/ca/ulaval/glo4003/domain/game/UnscheduledState.java
@@ -18,7 +18,7 @@ public class UnscheduledState implements GameScheduleState {
 	}
 
 	@Override
-	public void assignThisTicketToSchedule(Ticket ticket, int ticketNumber) {
+	public void assignThisTicketToSchedule(Ticket ticket, long ticketNumber) {
 
 	}
